+++ StepLang/Framework/IO/PrintFunction.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using StepLang.Framework.Conversion;
 using StepLang.Interpreting;
 using StepLang.Parsing.Expressions;
 
@@ -17,54 +18,12 @@ public class PrintFunction : NativeFunction
 
         var stringArgs = await arguments
             .EvaluateAsync(interpreter, cancellationToken)
-            .Select(RenderValue)
+            .Select(ToStringFunction.Render)
             .ToListAsync(cancellationToken);
 
         await Print(stdOut, string.Join("", stringArgs), cancellationToken);
 
         return VoidResult.Instance;
-    }
-
-    private static string RenderValue(ExpressionResult result)
-    {
-        switch (result)
-        {
-            case StringResult { Value: var stringValue }:
-                return stringValue;
-            case NumberResult { Value: var numberValue }:
-                return numberValue.ToString(CultureInfo.InvariantCulture);
-            case BoolResult { Value: var boolValue }:
-                return boolValue.ToString();
-            case NullResult:
-                return "null";
-            case ListResult { Value: var values }:
-                var renderedValues = values.Select(v =>
-                {
-                    string renderedValue;
-                    if (v is StringResult stringValue)
-                        renderedValue = $"\"{stringValue.Value}\"";
-                    else
-                        renderedValue = RenderValue(v);
-
-                    return renderedValue;
-                });
-                return $"[{string.Join(", ", renderedValues)}]";
-            case MapResult { Value: var pairs }:
-                var renderedPairs = pairs.Select(pair =>
-                {
-                    string renderedValue;
-                    if (pair.Value is StringResult stringValue)
-                        renderedValue = $"\"{stringValue.Value}\"";
-                    else
-                        renderedValue = RenderValue(pair.Value);
-
-                    return $"\"{pair.Key}\": {renderedValue}";
-                });
-                var mapContents = string.Join(", ", renderedPairs);
-                return $"{{{mapContents}}}";
-            default:
-                return $"[{result.ResultType}]";
-        }
     }
 
     protected virtual async Task Print(TextWriter output, string value, CancellationToken cancellationToken = default)
